improve performance of _align_array_to_sample
Always put the axis that we are aligning to as the first dimension
of the output array. This directly improves performance in the case
that axis=0, presumably because numpy does not have to deal with
fancy indexing, and should also improve performance in other cases
due to better cache locality.

New bottleneck is conversion of sparse to dense arrays