Avoid copying and densifying input data before training, use proper dataloaders instead (#81)
* Implement a custom dataloader and use it for training

This avoids the memory cost of converting the input data to dense
matrices beforehand. Currently only implemented for MuData input.

* allow passing a string to SmoothOptions.warp_groups

* fix preprocessing

scaling is calculated for the entire view, instead of per-feature

* fix r2 calculation

properly align factors with data

* fix saving with missing obs and string cols in metadata

* batch sampler go brrr

Increase speed by processing each minibatch as a whole. This avoids
having to subset the MuData object for each sample. Subsetting
MuData objects is a heavy operation due to the need to handle different
types of indices, aligning the modalities, etc.

New bottleneck is np.put_along_axis in _align_array_to_samples

* improve performance of _align_array_to_sample

Always put the axis that we are aligning to as the first dimension
of the output array. This directly improves performance in the case
that axis=0, presumably because numpy does not have to deal with
fancy indexing, and should also improve performance in other cases
due to better cache locality.

New bottleneck is conversion of sparse to dense arrays

* refactor in preparation for AnnDataDataset

* PRISMO: remove obsolete method

* document the PrismoDataset interface

* first implementation of AnnDataDictDataset

Basic training works, still needs perfromance optimization. Performance
optimization involves changing the API of apply: pass the unaligned
array to func, func can align itself if needed.

* API: use numpy arrays instead of lists

affects feature names, sample names, group names, view names

* correctly handle feature that are removed during preprocessing

* preprocessing: delete obsolete code

* make PrismoDataset act as a factory for its subclasses

* AnnDataDictDataset: subset minibatch before aligning

* properly align weights when calculating R2

* pass unaligned data matrices to the model

This avoids the cost of alignment (filling missing rows/cols with nans)
and shortens the time it takes to transfer data to the GPU. Instead,
indices of nonmissing samples/features are passed to the model, which
deals with missing data in Pyro plates.

* better handling of data transfers to GPU

* don't align features in apply(by_group=True)

* AnnDataDictDataset: use anndata.concat to concatenate views in apply

also use dask for lazy calculations when available

* AnnDataDictDataset: remove obsolete code

* better docs and API for array alignment

* add missing dependency

* port imputation to the new API

* MOFA compat: don't save imputed values

They are saved by mofapy2, but the MOFA R package does not read them.

* port MOFA compat saving

* simplify apply() code

* fix preprocessing

now identical to MOFA and earlier PRISMO

* port unit tests

* datasets: don't implement __getitem__

It's not well defined how to handle alignment/nonmissing
features/samples in that case.

* implement apply by view and port factor initialization to it

* core: allow disabling saving results after training

* fix Python 3.10 compatibility

* bump min. Python version to 3.11

* fix handling of nonmissing samples/features in the preprocessor

* AnnDataDictDataset: Correctly handle different orderings of features
in different views

* rework dataset alignment and handling of zero-variance features

In the previous implementation, nothing outside of the preprocessor
knew what features were removed, which led to issues with imputation
and R2 calculation. Also, the previous implementation could not handle
use_obs="intersection" or use_var="intersection". Both issues are now
solved.

* training loop: re-add single-batch special case

this should be much faster

* try harder to avoid copying arrays

* depend on development version of mudata

* improve readability a bit

* restructure PrismoDataSet docstring

* try to optimize dask chunksize for low memory